Archspec

Documentation for Archspec.

Base.getindex

Base.in

Base.getindex — Method

getindex(c::Compiler, v::VersionNumber)

Return the list of compiler flags for the given compiler version.

julia> Archspec.CPUTargets["x86_64_v2"].compilers["gcc"][v"5.2"]
"-march=x86_64_v2 -mtune=generic -mcx16 -msahf -mpopcnt -msse3 -msse4.1 -msse4.2 -mssse3"

julia> Archspec.CPUTargets["x86_64_v2"].compilers["gcc"][v"11.1"]
"-march=x86_64_v2 -mtune=generic"

Base.in — Method

in(feature::String, m::Microarchitecture) -> Bool
∈(feature::String, m::Microarchitecture) -> Bool

Return whether the given feature is available in the microarchitecture m.

julia> "sse3" in Archspec.CPUTargets["broadwell"]
true

julia> "avx512" in Archspec.CPUTargets["haswell"]
false